Make quadratic fixture accessible to all features test modules.
This is a useful test fixture for many features. Moving it to
conftest.py makes it so any submodule of tests.features can use it.
