+++ pvanalytics/tests/features/conftest.py
@@ -1,0 +1,15 @@
+"""Common fixtures for features tests."""
+import pytest
+import numpy as np
+import pandas as pd
+
+
+@pytest.fixture
+def quadratic():
+    """Downward facing quadratic.
+
+    Vertex at index 30 and roots at indices 0, 60.
+
+    """
+    q = -1000 * (np.linspace(-1, 1, 61) ** 2) + 1000
+    return pd.Series(q)
