+++ pvanalytics/tests/features/test_clipping.py
@@ -1,20 +1,8 @@
 """Tests for features.clipping"""
 import pytest
-import pandas as pd
 from pandas.util.testing import assert_series_equal
 import numpy as np
 from pvanalytics.features import clipping
-
-
-@pytest.fixture
-def quadratic():
-    """Downward facing quadratic.
-
-    Vertex at index 30 and roots at indices 0, 60.
-
-    """
-    q = -1000 * (np.linspace(-1, 1, 61) ** 2) + 1000
-    return pd.Series(q)
 
 
 @pytest.fixture
